+++ src/debugger/debugger.jl
@@ -1,9 +1,8 @@
-module Debugger
+isdebugging() = Debugger.isdebugging()
 
+module Debugger
 using ..Atom, MacroTools, ASTInterpreter2, Lazy, Hiccup
 
-# include("breakpoints.jl")
-include("stepper2.jl")
-# include("entry.jl")
+include("stepper.jl")
 
 end # module
